build: switch to next-like hierarchy
Create an app directory to contain components and hooks related to
business logic. Move index.html and main.tsx to a new folder. Modify
vite.config to support the new hierarchy. Create a "formatting" folder
in the app directory to contain formatting logic. Move the
InstructionsModal from the ui/components/* directory to the
app/pages/formatting/components directory, due to its relation to the
business logic. Delete src/components folder.
