--- src/App.tsx
@@ -1,11 +1,0 @@
-import FormattingForm from "@/components/formatting/formatting-form.tsx";
-import "./reset.css";
-function App() {
-    return (
-        <main>
-            <FormattingForm />
-        </main>
-    );
-}
-
-export default App;
