+++ src/app/pages/index.html
@@ -13,6 +13,6 @@
     </head>
     <body>
         <div id="root"></div>
-        <script type="module" src="/src/main.tsx"></script>
+        <script type="module" src="main.tsx"></script>
     </body>
 </html>
