+++ src/types/models/info-items.ts
@@ -1,6 +1,7 @@
 export type TInfoItemDescription =
     | string
     | { beforeBoldText: string; details: string }[];
+
 export type TInfoItem = {
     title: string;
     description: TInfoItemDescription;
