+++ src/ui/components/organisms/Modal/InstructionsModal/InstructionsModal.tsx
@@ -1,9 +1,6 @@
-import {
-    INSTRUCTION_ITEMS,
-    INSTRUCTIONS_TITLE,
-} from "@/configs/instructions.ts";
 import { InfoList } from "@/ui/components/organisms/InfoList";
 import { DefaultModal } from "@/ui/components/organisms/Modal/DefaultModal";
+import { INSTRUCTIONS_TITLE, INSTRUCTION_ITEMS } from "@/utils/consts.ts";
 import InfoIcon from "@mui/icons-material/Info";
 
 export const InstructionsModal = () => {
