+++ vite.config.ts
@@ -4,9 +4,17 @@ import tsconfigPaths from "vite-tsconfig-paths";
 // https://vite.dev/config/
 export default defineConfig({
     plugins: [react(), tsconfigPaths()],
+    root: "src/app/pages",
     base: "/commit-formatter/",
     server: {
         open: true,
         port: 3000,
+    },
+    build: {
+        rollupOptions: {
+            input: {
+                app: "src/app/pages/index.html",
+            },
+        },
     },
 });
